+++ src/validations/constants.ts
@@ -168,6 +168,7 @@ export const MEDIA_TYPES = ['IMAGE', 'VIDEO'];
 export const CONDITION_TYPE = ['HIDE', 'SHOW'];
 export const COMPARATORS = ['EQ'];
 export const MODAL_TYPE = ['MODEL_AND_PROPERTIES'];
+export const INHERIT_TYPES = ['PROPERTY', 'PROPERTY_LABEL'];
 
 export const CONFIGURATION_AS = [
   'BUTTONGROUP',